requirements.txt
aiogram==2.12.1
aiohttp==3.8.1
aioredis==1.3.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
environs==8.0.0
frozenlist==1.2.0
geojson==2.5.0
hiredis==2.0.0
idna==3.3
loguru==0.5.3
marshmallow==3.14.1
multidict==5.2.0
pyowm==3.2.0
PySocks==1.7.1
python-dotenv==0.19.2
python-telegram-bot-pagination==0.0.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
xmltodict==0.12.0
yarl==1.7.2
